refactor: move byte to tensor to image bytes
