+++ docarray/typing/tensor/image/abstract_image_tensor.py
@@ -6,17 +6,22 @@ from docarray.typing.tensor.abstract_tensor import AbstractTensor
 
 class AbstractImageTensor(AbstractTensor, ABC):
     @abstractmethod
-    def to_bytes(self) -> bytes:
+    def to_bytes(self, format: str = 'PNG') -> bytes:
         """
         Convert image tensor to bytes.
         """
         from PIL import Image
 
+        if format == 'jpg':
+            format = 'jpeg'  # unify it to ISO standard
+
         tensor = self.get_comp_backend().to_numpy(self)
-        pil_image = Image.fromarray(tensor)
+
+        mode = 'RGB' if tensor.ndim == 3 else 'L'
+        pil_image = Image.fromarray(tensor, mode=mode)
 
         with io.BytesIO() as buffer:
-            pil_image.save(buffer, format='PNG')
+            pil_image.save(buffer, format=format)
             img_byte_arr = buffer.getvalue()
 
         return img_byte_arr
