+++ docarray/typing/url/image_url.py
@@ -1,5 +1,3 @@
-import io
-import struct
 from typing import TYPE_CHECKING, Any, Optional, Tuple, Type, TypeVar, Union
 
 import numpy as np
@@ -8,7 +6,6 @@ from docarray.typing.proto_register import _register_proto
 from docarray.typing.url.any_url import AnyUrl
 
 if TYPE_CHECKING:
-    import PIL
     from pydantic import BaseConfig
     from pydantic.fields import ModelField
 
@@ -91,122 +88,3 @@ class ImageUrl(AnyUrl):
 
         buffer = ImageBytes(self.load_bytes(timeout=timeout))
         return buffer.load(width, height, axis_layout)
-
-
-def _image_tensor_to_bytes(arr: np.ndarray, image_format: str) -> bytes:
-    """
-    Convert image-ndarray to buffer bytes.
-
-    :param arr: Data representations of the png.
-    :param image_format: `png` or `jpeg`
-    :return: Png in buffer bytes.
-    """
-
-    if image_format not in IMAGE_FILE_FORMATS:
-        raise ValueError(
-            f'image_format must be one of {IMAGE_FILE_FORMATS},'
-            f'receiving `{image_format}`'
-        )
-    if image_format == 'jpg':
-        image_format = 'jpeg'  # unify it to ISO standard
-
-    arr = arr.astype(np.uint8).squeeze()
-
-    if arr.ndim == 1:
-        # note this should be only used for MNIST/FashionMNIST dataset,
-        # because of the nature of these two datasets
-        # no other image data should flattened into 1-dim array.
-        image_bytes = _png_to_buffer_1d(arr, 28, 28)
-    elif arr.ndim == 2:
-        from PIL import Image
-
-        im = Image.fromarray(arr).convert('L')
-        image_bytes = _pillow_image_to_buffer(im, image_format=image_format.upper())
-    elif arr.ndim == 3:
-        from PIL import Image
-
-        im = Image.fromarray(arr).convert('RGB')
-        image_bytes = _pillow_image_to_buffer(im, image_format=image_format.upper())
-    else:
-        raise ValueError(
-            f'{arr.shape} ndarray can not be converted into an image buffer.'
-        )
-
-    return image_bytes
-
-
-def _png_to_buffer_1d(arr: np.ndarray, width: int, height: int) -> bytes:
-    import zlib
-
-    pixels = []
-    for p in arr[::-1]:
-        pixels.extend([p, p, p, 255])
-    buf = bytearray(pixels)
-
-    # reverse the vertical line order and add null bytes at the start
-    width_byte_4 = width * 4
-    raw_data = b''.join(
-        b'\x00' + buf[span : span + width_byte_4]
-        for span in range((height - 1) * width_byte_4, -1, -width_byte_4)
-    )
-
-    def png_pack(png_tag, data):
-        chunk_head = png_tag + data
-        return (
-            struct.pack('!I', len(data))
-            + chunk_head
-            + struct.pack('!I', 0xFFFFFFFF & zlib.crc32(chunk_head))
-        )
-
-    png_bytes = b''.join(
-        [
-            b'\x89PNG\r\n\x1a\n',
-            png_pack(b'IHDR', struct.pack('!2I5B', width, height, 8, 6, 0, 0, 0)),
-            png_pack(b'IDAT', zlib.compress(raw_data, 9)),
-            png_pack(b'IEND', b''),
-        ]
-    )
-
-    return png_bytes
-
-
-def _pillow_image_to_buffer(image: 'PIL.Image.Image', image_format: str) -> bytes:
-    img_byte_arr = io.BytesIO()
-    image.save(img_byte_arr, format=image_format)
-    img_bytes = img_byte_arr.getvalue()
-    return img_bytes
-
-
-def _move_channel_axis(
-    tensor: np.ndarray, axis_layout: Tuple[str, str, str] = ('H', 'W', 'C')
-) -> np.ndarray:
-    """Moves channel axis around."""
-    channel_to_offset = {'H': 0, 'W': 1, 'C': 2}
-    permutation = tuple(channel_to_offset[axis] for axis in axis_layout)
-    return np.transpose(tensor, permutation)
-
-
-def _to_image_tensor(
-    source: Union[str, bytes, io.BytesIO],
-    width: Optional[int] = None,
-    height: Optional[int] = None,
-) -> 'np.ndarray':
-    """
-    Convert an image blob to tensor
-
-    :param source: binary blob or file path
-    :param width: the width of the image tensor.
-    :param height: the height of the tensor.
-    :return: image tensor
-    """
-    from PIL import Image as PILImage
-
-    raw_img = PILImage.open(source)
-    if width or height:
-        new_width = width or raw_img.width
-        new_height = height or raw_img.height
-        raw_img = raw_img.resize((new_width, new_height))
-    try:
-        return np.array(raw_img.convert('RGB'))
-    except Exception:
-        return np.array(raw_img)
